# [NOMERGE] 49% faster pattern matcher (passes process replay)