--- a/tinygrad/codegen/uops.py
+++ b/tinygrad/codegen/uops.py
@@ -148,7 +148,7 @@ class PatternMatcher:
       else:
         self.pdict[(p.op, p.arg)].append((p, fxn))
 
-  @functools.lru_cache(None)
+
   def rewrite(self, uop:UOp) -> Optional[UOp]:
     for p,fxn in itertools.chain(self.pdict[(uop.op, uop.arg)], self.pdict[(uop.op, None)]):
       store: Dict[str, UOp] = {}