--- a/tinygrad/codegen/uops.py
+++ b/tinygrad/codegen/uops.py
@@ -31,7 +31,7 @@ def ufix(dtype: Optional[DType], x): return UOp.const(dtype, x) if not isinstanc
 class UOp:
   op: UOps
   dtype: Optional[DType] = None
-  src: Tuple[UOp, ...] = ()
+  src: Tuple[UOp, ...] = tuple()
   arg: Any = None
   def commutative(self) -> bool:
     return self.op is UOps.ALU and self.arg in {BinaryOps.ADD, BinaryOps.MUL, BinaryOps.MAX, BinaryOps.CMPNE, BinaryOps.XOR}
@@ -64,7 +64,6 @@ class UOp:
     if isinstance(b, Variable): return UOp(UOps.DEFINE_VAR, dtype, (), b)
     return UOp(UOps.CONST, dtype, arg=dtypes.as_const(b, dtype) if dtype is not None else b)
   @staticmethod
-  @functools.lru_cache(None)
   def alu(arg, *src:UOp): return UOp(UOps.ALU, dtypes.bool if arg in {BinaryOps.CMPLT, BinaryOps.CMPNE} else src[-1].dtype, src, arg)
   @staticmethod
   def load(*src:UOp, dtype:Optional[DType]=None, **kwargs): return UOp(UOps.LOAD, dtype, tuple(src)+tuple(kwargs.values()))
@@ -91,37 +90,29 @@ def uop_alu_resolve(u:UOp) -> sint:
 
 # *** simplification logic ***
 
-@dataclass
+@dataclass(frozen=True)
 class UPat:
-  op: Any = () # should be Tuple[UOps, ...]]
+  op: Tuple[UOps, ...] = ()
   arg: Any = None
-  src: Any = None  # should be Optional[Tuple[UOps, ...]]], None is not the same as tuple()!
+  src: Optional[Union[List[UPat], Tuple[UPat, ...], UPat]] = None
   name: Optional[str] = None
-  dtype: Any = ()  # should be Optional[Tuple[UOps, ...]]]
-  allow_len: Union[Set[int], Tuple[int,...]] = ()  # should be Tuple[int]
-
-  def __post_init__(self):
-    if isinstance(self.op, UOps): self.op = (self.op,)
-    if isinstance(self.dtype, DType): self.dtype = (self.dtype,)
-    elif isinstance(self.dtype, set): self.dtype = tuple(self.dtype)
-    if isinstance(self.allow_len, set): self.allow_len = tuple(self.allow_len)
-    # only one if it's a tuple
-    # try all permutations if it's a list
-    # repeat if it's a UPat
-    self.src = tuple(itertools.permutations(self.src)) if isinstance(self.src,list) else (self.src,) if isinstance(self.src,tuple) else self.src
+  dtype: Tuple[DType, ...] = ()
+  allow_len: Tuple[int,...] = ()
 
   @staticmethod
   def compile(u: UOp, name:Optional[str]=None) -> UPat:
-    if u.op is UOps.VAR: return UPat(name=name or u.arg, dtype=u.dtype) if len(u.src) == 0 else UPat.compile(u.src[0], name or u.arg)
-    return UPat(u.op, u.arg, (list if u.commutative() else tuple)([UPat.compile(src) for src in u.src]) if u.src != () else None, name, u.dtype)
-
-T = TypeVar("T")
-def __unmatch(m1:Union[T, Set[T]], m2:T) -> bool: return m2 not in m1 if isinstance(m1, set) else m2 != m1
+    if u.op is UOps.VAR:
+      return UPat(name=name or u.arg, dtype=(u.dtype,) if u.dtype is not None else ()) if len(u.src) == 0 else UPat.compile(u.src[0], name or u.arg)
+    op, dtype = (u.op,) if u.op is not None else (), (u.dtype,) if u.dtype is not None else (),
+    return UPat(op, u.arg, (list if u.commutative() else tuple)([UPat.compile(src) for src in u.src]) if u.src != () else None, name, dtype)
 
 def _match(uop:UOp, pat:UPat, store:Dict[str, UOp]) -> bool:
   if pat.op and uop.op not in pat.op: return False
+  if pat.dtype and uop.dtype and uop.dtype not in pat.dtype: return False
+  if pat.arg is not None and (uop.arg not in pat.arg if isinstance(pat.arg, (set,tuple)) else pat.arg != uop.arg): return False
   if pat.src:
-    for vp in pat.src if isinstance(pat.src, tuple) else [(pat.src,)*len(uop.src)]:
+    for vp in itertools.permutations(pat.src) if isinstance(pat.src, list) else (pat.src,) if isinstance(pat.src,tuple) \
+      else [(pat.src,)*len(uop.src)]:
       if len(uop.src) != len(vp) and (len(uop.src) not in pat.allow_len): return False
       new_store = store.copy()
       for uu, vv in zip(uop.src, vp):
@@ -130,9 +121,7 @@ def _match(uop:UOp, pat:UPat, store:Dict[str, UOp]) -> bool:
         store.update(new_store)
         break
     else: return False  # None of the src matches
-  if pat.dtype and uop.dtype and uop.dtype not in pat.dtype: return False
   if pat.name is not None and store.setdefault(pat.name,uop) is not uop: return False
-  if pat.arg is not None and (uop.arg not in pat.arg if isinstance(pat.arg, (set,tuple)) else pat.arg != uop.arg): return False
   return True
 
 class PatternMatcher:
@@ -143,10 +132,7 @@ class PatternMatcher:
     for p,fxn in self.patterns:
       if isinstance(p, UOp): p = UPat.compile(p)
       assert p.op is not None
-      if isinstance(p.op, (set, tuple)):
-        for uop in p.op: self.pdict[(uop, p.arg)].append((p, fxn))
-      else:
-        self.pdict[(p.op, p.arg)].append((p, fxn))
+      for uop in p.op: self.pdict[(uop, p.arg)].append((p, fxn))
 
   def rewrite(self, uop:UOp) -> Optional[UOp]:
     for p,fxn in itertools.chain(self.pdict[(uop.op, uop.arg)], self.pdict[(uop.op, None)]):
@@ -174,37 +160,37 @@ def loop_collapse(loop_start, loop_end, compval, idx, mval, multconst, rng):
 # this is symbolic 2.0
 constant_folder = PatternMatcher([
   # arange loop folding (early)
-  (UPat(UOps.ALU, TernaryOps.WHERE, src=(UPat(UOps.ALU, BinaryOps.CMPLT, src=(
-    UPat(UOps.ALU, BinaryOps.ADD, src=[UPat(name="idx"), UPat(UOps.ALU, BinaryOps.MUL, src=[UPat(UOps.CONST, name="mval"),
-      UPat(UOps.RANGE, name="rng", src=(UPat(name="loop_start"), UPat(name="loop_end")))])]),
-      UPat(UOps.CONST, name="compval"))), UPat(UOps.CONST, name="multconst"), UPat(UOps.CONST, 0))), loop_collapse),
-  (UPat(UOps.ALU, TernaryOps.WHERE, src=(UPat(UOps.ALU, BinaryOps.CMPLT, src=(
-    UPat(UOps.ALU, BinaryOps.ADD, src=[UPat(name="idx"), UPat(UOps.ALU, UnaryOps.NEG, src=[
-      UPat(UOps.RANGE, name="rng", src=(UPat(name="loop_start"), UPat(name="loop_end")))])]),
-      UPat(UOps.CONST, name="compval"))), UPat(UOps.CONST, name="multconst"), UPat(UOps.CONST, 0))),
+  (UPat((UOps.ALU,), TernaryOps.WHERE, src=(UPat((UOps.ALU,), BinaryOps.CMPLT, src=(
+    UPat((UOps.ALU,), BinaryOps.ADD, src=[UPat(name="idx"), UPat((UOps.ALU,), BinaryOps.MUL, src=[UPat((UOps.CONST,), name="mval"),
+      UPat((UOps.RANGE,), name="rng", src=(UPat(name="loop_start"), UPat(name="loop_end")))])]),
+      UPat((UOps.CONST,), name="compval"))), UPat((UOps.CONST,), name="multconst"), UPat((UOps.CONST,), 0))), loop_collapse),
+  (UPat((UOps.ALU,), TernaryOps.WHERE, src=(UPat((UOps.ALU,), BinaryOps.CMPLT, src=(
+    UPat((UOps.ALU,), BinaryOps.ADD, src=[UPat(name="idx"), UPat((UOps.ALU,), UnaryOps.NEG, src=[
+      UPat((UOps.RANGE,), name="rng", src=(UPat(name="loop_start"), UPat(name="loop_end")))])]),
+      UPat((UOps.CONST,), name="compval"))), UPat((UOps.CONST,), name="multconst"), UPat((UOps.CONST,), 0))),
       lambda **kwargs: loop_collapse(mval=UOp.const(dtypes.int, -1), **kwargs)),
   # sum collapse to mul (with possible GEP)
-  (UPat(UOps.PHI, src=(UPat(UOps.DEFINE_ACC, name="phi_input", src=[UPat(UOps.CONST), UPat(UOps.RANGE, name="loop")]),
-                       UPat(UOps.ALU, BinaryOps.ADD, src=(UPat(name="val1"), UPat(name="val2"))))), sum_collapse),
-  (UPat(UOps.PHI, src=(UPat(UOps.GEP, name="phi_input", src=(UPat(UOps.DEFINE_ACC, src=[UPat(UOps.CONST), UPat(UOps.RANGE, name="loop")]),)),
-                       UPat(UOps.ALU, BinaryOps.ADD, src=(UPat(name="val1"), UPat(name="val2"))))), sum_collapse),
+  (UPat((UOps.PHI,), src=(UPat((UOps.DEFINE_ACC,), name="phi_input",src=[UPat((UOps.CONST,)),
+   UPat((UOps.RANGE,), name="loop")]), UPat((UOps.ALU,), BinaryOps.ADD, src=(UPat(name="val1"), UPat(name="val2"))))), sum_collapse),
+  (UPat((UOps.PHI,), src=(UPat((UOps.GEP,), name="phi_input",src=(UPat((UOps.DEFINE_ACC,), src=[UPat((UOps.CONST,)),
+   UPat((UOps.RANGE,), name="loop")]),)), UPat((UOps.ALU,), BinaryOps.ADD, src=(UPat(name="val1"), UPat(name="val2"))))), sum_collapse),
   # deal with UNMUL
-  (UPat(UOps.ALU, BinaryOps.MUL, [UPat(UOps.CONST, name="c1"), UPat(UOps.UNMUL, src=[UPat(UOps.CONST, name="c2"), UPat(name="v")])]),
+  (UPat((UOps.ALU,), BinaryOps.MUL, [UPat((UOps.CONST,), name="c1"), UPat((UOps.UNMUL,), src=[UPat((UOps.CONST,), name="c2"), UPat(name="v")])]),
    lambda c1,c2,v: v if c1.arg == c2.arg else None),
   (UOp(UOps.UNMUL, src=(UOp.const(None, 0).name('zero'), UOp.var())), lambda zero: zero),
   (UOp(UOps.UNMUL).name('unmul').cast().name('root'), lambda root,unmul: UOp(UOps.UNMUL, root.dtype, (unmul.src[0].cast(root.dtype), unmul.src[1]))),
   # max on special can go away (TODO: special should be variable, same thing applies)
   (UOp.max(UOp.cvar('c'), UOp(UOps.SPECIAL).name('s')), lambda c,s: c if (s.arg[2]-1) <= c.arg else None),
   # const rules
-  (UPat(UOps.GEP, name="root", src=(UPat(UOps.CONST, name="c"),)), lambda root, c: UOp.const(root.dtype, c.arg)),
-  (UPat(UOps.CAST, name="root", src=UPat(UOps.CONST, name="c")), lambda root, c: UOp.const(root.dtype, c.arg)),
+  (UPat((UOps.GEP,), name="root", src=(UPat((UOps.CONST,), name="c"),)), lambda root, c: UOp.const(root.dtype, c.arg)),
+  (UPat((UOps.CAST,), name="root", src=UPat((UOps.CONST,), name="c")), lambda root, c: UOp.const(root.dtype, c.arg)),
   # a phi on a DEFINE_ACC without loops or a CONST is a noop. this is for correctness, not just speed
-  (UPat(UOps.PHI, src=(UPat(UOps.DEFINE_ACC, name="acc"), UPat(name="acc"))), lambda acc: UOp.cast(acc.src[0], acc.dtype)),
-  (UPat(UOps.PHI, src=(UPat(UOps.DEFINE_ACC, src=(UPat(UOps.CONST),)), UPat(name="x"))), lambda x: x),
-  (UPat(UOps.PHI, src=(UPat(UOps.CONST), UPat(name="x"))), lambda x: x),
+  (UPat((UOps.PHI,), src=(UPat((UOps.DEFINE_ACC,), name="acc"), UPat(name="acc"))), lambda acc: UOp.cast(acc.src[0], acc.dtype)),
+  (UPat((UOps.PHI,), src=(UPat((UOps.DEFINE_ACC,), src=(UPat((UOps.CONST,)),)), UPat(name="x"))), lambda x: x),
+  (UPat((UOps.PHI,), src=(UPat((UOps.CONST,)), UPat(name="x"))), lambda x: x),
   # a DEFINE_ACC without inputs is a const + GEP on a const is the const
-  (UPat(UOps.DEFINE_ACC, name="root", src=(UPat(UOps.CONST),)), lambda root: UOp.cast(root.src[0], root.dtype)),
-  (UPat(UOps.GEP, name="root", src=(UPat(UOps.CONST, name="x"),)), lambda root,x: UOp.const(root.dtype, x.arg)),
+  (UPat((UOps.DEFINE_ACC,), name="root", src=(UPat((UOps.CONST,)),)), lambda root: UOp.cast(root.src[0], root.dtype)),
+  (UPat((UOps.GEP,), name="root", src=(UPat((UOps.CONST,), name="x"),)), lambda root,x: UOp.const(root.dtype, x.arg)),
   # max -2147483648
   (UOp.max(UOp.var('x'), UOp.const(dtypes.int, -2147483648)), lambda x: x),
   # bool < False is always false, True < bool is always false
@@ -214,7 +200,8 @@ constant_folder = PatternMatcher([
   (UOp.alu(TernaryOps.WHERE, UOp.var(), UOp.var("val"), UOp.var("val")), lambda val: val),
   (UOp.alu(TernaryOps.WHERE, UOp.cvar('gate'), UOp.var('c0'), UOp.var('c1')), lambda gate, c0, c1: c0 if gate.arg else c1),
   # ** constant folding **
-  (UPat(UOps.ALU, name="root", src=UPat(UOps.CONST)), lambda root: UOp.const(root.dtype, exec_alu(root.arg, root.dtype, [x.arg for x in root.src]))),
+  (UPat((UOps.ALU,), name="root", src=UPat((UOps.CONST,))),
+   lambda root: UOp.const(root.dtype, exec_alu(root.arg, root.dtype, [x.arg for x in root.src]))),
   # ** self folding **
   (-(-UOp.var('x')), lambda x: x),    # -(-x) -> x
   (UOp.var('x') + 0, lambda x: x),    # x+0 -> x
@@ -267,14 +254,15 @@ constant_folder = PatternMatcher([
   (UOp.store(UOp.var("buf"), UOp.var("idx"), UOp(UOps.CAST, src=tuple(UOp(UOps.GEP, arg=i, src=(UOp.var("val"),)) for i in range(4)))), UOp.store),
   (UOp.store(UOp.var("buf"), UOp.var("idx"), UOp(UOps.CAST, src=tuple(UOp(UOps.GEP, arg=i, src=(UOp.var("val"),)) for i in range(2)))), UOp.store),
   # CAST-PHI-GEP -> PHI-CAST
-  (UPat(UOps.CAST, name="root", src=tuple(UPat(UOps.PHI, src=(UPat(UOps.GEP, i, src=(UPat(name="val"),)), UPat(name=f"v{i}"))) for i in range(4))),
-    lambda root, val, v0, v1, v2, v3: UOp(UOps.PHI, root.dtype, (val, UOp(UOps.CAST, val.dtype, (v0, v1, v2, v3))))),
-  (UPat(UOps.CAST, name="root", src=tuple(UPat(UOps.PHI, src=(UPat(UOps.GEP, i, src=(UPat(name="val"),)), UPat(name=f"v{i}"))) for i in range(2))),
-    lambda root, val, v0, v1: UOp(UOps.PHI, root.dtype, (val, UOp(UOps.CAST, val.dtype, (v0, v1))))),
+  (UPat((UOps.CAST,), name="root", src=tuple(UPat((UOps.PHI,), src=(UPat((UOps.GEP,), i, src=(UPat(name="val"),)),
+   UPat(name=f"v{i}"))) for i in range(4))), lambda root, val, v0, v1, v2, v3: UOp(UOps.PHI, root.dtype,
+     (val, UOp(UOps.CAST, val.dtype, (v0, v1, v2, v3))))),
+  (UPat((UOps.CAST,), name="root", src=tuple(UPat((UOps.PHI,), src=(UPat((UOps.GEP,), i, src=(UPat(name="val"),)),
+   UPat(name=f"v{i}"))) for i in range(2))), lambda root, val, v0, v1: UOp(UOps.PHI, root.dtype, (val, UOp(UOps.CAST, val.dtype, (v0, v1))))),
   # NEG/CMPLT -> CMPLT
   (UOp.lt(-UOp.var('x'), UOp.cvar('c', dtypes.int)), lambda c,x: UOp.lt(UOp.const(c.dtype, -c.arg), x)),
   # cast NOOP (NOTE: it's str to deal with PtrDType)
-  (UPat(UOps.CAST, name="root"), lambda root: root.src[0] if str(root.dtype) == str(root.src[0].dtype) else None),
+  (UPat((UOps.CAST,), name="root"), lambda root: root.src[0] if str(root.dtype) == str(root.src[0].dtype) else None),
   # fold gated LOAD/STORE
   (UOp.load(UOp.var("buf"), UOp.var("idx"), UOp.const(None, 1), UOp.cvar("var")), lambda buf,idx,var: UOp.load(buf, idx, dtype=var.dtype)),
   (UOp.load(UOp.var("buf"), UOp.var("idx"), UOp.const(None, 1), UOp.cvar("var"), UOp.var("barrier")),
@@ -284,7 +272,7 @@ constant_folder = PatternMatcher([
   (UOp.store(UOp.var("buf"), UOp.var("idx"), UOp.var("val"), UOp.const(None, 1)), UOp.store),
   (UOp.store(UOp.var(), UOp.var(), UOp.var(), UOp.const(None, 0)), lambda: UOp(UOps.NOOP)),
   # remove NOOPs from SINK
-  (UPat(UOps.SINK, name="root"),
+  (UPat((UOps.SINK,), name="root"),
     lambda root: UOp(UOps.SINK, root.dtype, a, root.arg) if len(a:=tuple(x for x in root.src if x.op is not UOps.NOOP)) != len(root.src) else None)
 ])