… test/test_pattern_matcher.py
… tinygrad/renderer/assembly.py