Some work towards: https://github.com/tinygrad/tinygrad/issues/4878. Should now be at~60% of the baseline.

* Caches on UOp methods
* No unhashable members of UPat (set, list -> tuple), create permutations on pat construction
* Replace `__unmatch` with `in` check
* Check op and src first in `_match`
* pull cached rewrite function out of double while loop

Once all usages of UPat are changed, most of the `__post_int__` can go and the typing can be fixed again to simpler tyes (`Optional[Union[DType, set(DType]]` -> `Tuple[DType,...]`. I held off for now because this would make the diff explode.


## Timing 
759.37/509.61 = 1.49

`PROFILE=0 python -O test/external/external_benchmark_schedule.py` best of 5

master
```
***** model forward in  15.84 ms
***** model schedule in   5.33 ms
***** model lower in 759.37 ms
```
branch
```
***** model forward in  15.84 ms
***** model schedule in   5.35 ms
***** model lower in 509.61 ms
```