@DKormann 2024-06-26
caching the staticmethods of UOp ist interesting. will that create a permanent persistent caching?