@roelofvandijk 2024-06-26
`Permanent persistent` for as long as the python execution runs. On staticmethods it won't store copies of the class instances.