@chenyuxyz 2024-06-26
i think the goal of the bounty is to improve the algorithm, not putting the current algorithm behind caches and make immutable objects mutable