@geohot 2024-06-26
Caches like that on ALU are dangerous. They will keep all parent UOps around forever.