@geohot 2024-06-26
In general, I'm opposed to caches on any UOp with parents. They will make them live the whole lifetime of the Python process.