@roelofvandijk 2024-06-26
> Caches like that on ALU are dangerous. They will keep all parent UOps around forever. Caching const is probably okay.

I added a cache clear after the rewrite. Is that still too dangerous (in case the cache clear is removed accidentally)?

> I see some stuff in here that's wins and other stuff that's trading complexity for speed. 

As is tradition. I'll try to stick to more of the first.

> I don't like the loss of type specificity on UPat. 

That is not the end state of this PR, if I modify all the instantiations to have the correct types right away, the types come back (and will be more specific e.g. `Optional[Union[DType, set(DType]]` becomes `Tuple[DType,...]`). Diff will explode.

> Is `in` with a tuple of size 1 fast?

Faster (and simpler) than an `isinstance` check in `__unmatch`. My idea here was to define `UPat.ops` and `UPat.dtypes` which are tuples of matching `Uops` and `Dtypes`. Did not do the rename to keep the diff small. Moreover, tuples are immutable and hashable (and sets are not).

> And why are the other checks moved below the recursive src check?

Statistical optimization based on the benchmark. Ops unmatch more often, then src (with early stopping!), etc. I also found this way simpler: it's a match, unless somewhere it gets unmatched.

```
Total time: 0.584024 s
File: /Users/rvd/src/rvd/tinygrad/tinygrad/codegen/uops.py
Function: _match at line 119

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   119                                           @profile
   120                                           def _match(uop:UOp, pat:UPat, store:Dict[str, UOp]) -> bool:
   121    473152      68088.0      0.1     11.7    if pat.op and uop.op not in pat.op: 
   122    179127      21690.0      0.1      3.7      return False
   123    294025      35267.0      0.1      6.0    if pat.src:
   124    425878      56474.0      0.1      9.7      for vp in pat.src if isinstance(pat.src, tuple) else [(pat.src,)*len(uop.src)]:
   125    286733      44834.0      0.2      7.7        if len(uop.src) != len(vp) and (len(uop.src) not in pat.allow_len): 
   126     37310       8040.0      0.2      1.4          return False
   127    249423      25025.0      0.1      4.3        new_store = store.copy()
   128    349887      63706.0      0.2     10.9        for uu, vv in zip(uop.src, vp):
   129    346318     134695.0      0.4     23.1          if not _match(uu, vv, new_store): break
   130                                                 else:
   131      3569        616.0      0.2      0.1          store.update(new_store)
   132      3569        283.0      0.1      0.0          break
   133    139145      33606.0      0.2      5.8      else: return False  # None of the src matches
   134    117570      17435.0      0.1      3.0    if pat.dtype and uop.dtype and uop.dtype not in pat.dtype: 
   135      2289        643.0      0.3      0.1      return False
   136    115281      20182.0      0.2      3.5    if pat.name is not None and store.setdefault(pat.name,uop) is not uop: 
   137      4860       1445.0      0.3      0.2      return False
   138    110421      14737.0      0.1      2.5    if pat.arg is not None and (uop.arg not in pat.arg if isinstance(pat.arg, (set,tuple)) else pat.arg != uop.arg): 
   139      9248       3155.0      0.3      0.5      return False
   140    101173      34103.0      0.3      5.8    return True
```