@roelofvandijk 2024-06-26
I'm leaving this up, but don't expect it to go in. I still believe the all-tuples UPat idea has merit, but it does make the UPat definitions uglier.