@geohot 2024-06-28
How much is the reordering in match vs the UPat change? I do think the syntax looks worse now.

Before:
```
diane@blak:~/tinygrad$ PROFILE=0 python3 test/external/external_benchmark_schedule.py
***** model forward in  19.66 ms
***** model schedule in   7.47 ms
***** model lower in 798.28 ms
```

After:
```
diane@blak:~/tinygrad$ PROFILE=0 python3 test/external/external_benchmark_schedule.py
***** model forward in  20.15 ms
***** model schedule in   7.10 ms
***** model lower in 740.92 ms
```